Avoid "stuck dim LED" effect
During measurement, do not enable PWM output already.  Otherwise, with
OCR0A = 0, it resulted in a one-clock-cycle (7 µs) turn-on period of
the LED already, causing it to be dim.

Abort current measurement if measurement time exceeds the period
defined for full darkness; this avoids the measurement taking
virtually forever, so the next measurement was already triggered 10
seconds later, restarting everything from the beginning. This
situation led up to a dim LED (due to the PWM pulse mentioned before)
that never got bright when light was flipped from brightness to
darkness.